Now drawing with a very simple shader hard-coded into the source. Eventually we will load shaders from text files. (One unified .glsl file or a .frag and .vert for the vertex and fragment shaders respectively.)
NOTE: libs is the folder where any third-party libraries (that are not compiled) will go. GLee is an openGL extension loader library that lets us check for device capabilities pretty easily and do lots of other stuff we don't know how to do yet. It's completely cross-platform and DO NOT move the position of it's #include in GameApp.h. It must be included before the openGL headers.